D14 P2 longs

+++ src/year_2019/day14/Day14.java
@@ -5,7 +5,7 @@ import java.io.IOException;
 public class Day14 {
 
 
-    public static Integer part1(String fileName) throws IOException {
+    public static Long part1(String fileName) throws IOException {
         ReactionInfo reactionInfo = new ReactionInfo(fileName);
         return reactionInfo.findLeastRequiredOreForOneFuel();
     }
